backend/requirements.txt
openicl
python-dotenv
flask
flasgger
importlib
flask-cors
accuracy
sacrebleu
rouge
squad
numpy